+++ lib/pactas_itero/api/customers.rb
@@ -1,7 +1,7 @@
 module PactasItero
   module Api
     module Customers
-      def create_customer(options= {})
+      def create_customer(options = {})
         options = options.camelize_keys
         post "api/v1/customers", options
       end
@@ -9,6 +9,11 @@ module PactasItero
       def customers(options = {})
         options = options.camelize_keys
         get "api/v1/customers", options
+      end
+
+      def customer(customer_id, options = {})
+        options = options.camelize_keys
+        get "api/v1/customers/#{customer_id}", options
       end
 
       def update_customer(customer_id, options = {})