+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [UNRELEASED]
 ### Added
 - get a self service token for a given contract
+- get a specific customer by id
 
 ### Changed
 - update version of jruby to 9.1.5.0